+++ core/__init__.py
@@ -1,5 +1,6 @@
 """mold-flow-fangate core package."""
 
+from .fan_gate import FanGatePlateConfig, build_fan_gate_plate_geometry
 from .fill_player import (
     CONTROLS_HEIGHT_PX,
     build_fill_player_html,
@@ -31,6 +32,8 @@ __all__ = [
     "cross_wlf_viscosity",
     "Geometry",
     "build_demo_geometry",
+    "FanGatePlateConfig",
+    "build_fan_gate_plate_geometry",
     "HeleShawSolver",
     "MultilayerHeleShawSolver",
     "MultilayerFlowResult",